Add files via upload
A port scanner is an application which is made to probe a host or server to identify open ports. Bad actors can use port scanners to exploit vulnerabilities by finding network services running on a host. They can also be used by security analysts to confirm network security policies. Running a port scan on a network or server reveals which ports are open and listening (receiving information) as well as revealing the presence of security devices, such as firewalls, that are present between the sender and the target. This technique is known as fingerprinting.

It is also valuable for testing network security and the strength of the system’s firewall. Due to this functionality, it is also a popular reconnaissance tool for attackers seeking a weak point of access to break into a computer.

Ports vary in their services offered. They are numbered from 0 to 65535, but certain ranges are more frequently used. Ports 0 to 1023 are identified as the “well-known ports” or standard ports and have been assigned services by the Internet Assigned Numbers Authority (IANA). Some of the most prominent ports and their assigned services include:

Port 20 (UDP) — File Transfer Protocol (FTP) for data transfer Port 22 (TCP) — Secure Shell (SSH) protocol for secure logins, FTP, and port forwarding Port 23 (TCP) — Telnet protocol for unencrypted text commutations Port 53 (UDP) — Domain Name System (DNS) translates names of all computers on internet-to-IP addresses Port 80 (TCP) — World Wide Web HTTP There are standard services offered on ports after 1023 as well and ports that, if open, indicate an infected system due to its popularity with some far-reaching Trojans and viruses.